Add extra column to align the change arrow
This is untested, but I *think* this will help align the change column now that BGG changed their arrow to an icon and it no longer respects the monospace of the table font. The intent is to just put that icon into an additional unlabeled column. It might not be the best looking solution but I figured I'd try my hand at contributing. No worries if you want to use a different approach. Thanks for the weekly list!